Allow to use alternative font on small font size
On smaller font size, it is sometimes better to switch to a bitmap font
automatically.

This change allows two fonts to be configured - one for larger sizes and
another for smaller sizes. An optional 'small_font' configuration is
possible, with an 'upper_bound' on the size which determines when to use
that font, based on currently used size.

For example, I with this commit I use Hack, a TTF antialised font for
large sizes, and Terminus, a bitmap font, for small sizes:

font:
  normal:
    family: Hack
  bold:
    family: Hack
  italic:
    family: Hack
  size: 13.0

small_font:
  upper_bound:
      size: 11.0
  font:
      normal:
        family: Terminus
      bold:
        family: Terminus
      italic:
        family: Terminus
      size: 16.0